Set max wait time if setVoicemailNumber is stuck
In some rare cases, Google can be in the middle of a
switch that happens right after the phone boots. Then
at the same time, ongoing configuration changes that
require interaction with the SIM can get stuck until
the RIL request times out. In order to limit the amount
of time that the binder thread blocks, set a hard
deadline of 2 seconds. The potential side effect would
be that an extremely slow SIM card access might result
in a false negative response, but this should not happen
in practice. If it does, still preferable to avoid the
ANR.

Bug: 29190397
Test: es posible
Flag: EXEMPT bugfix
Change-Id: Idd3ebf510f9422e3dfbdb308fafa4756c0b7c17e