Add notification when satellite availability changes
When satellite availaibility changes, we need to notify the state to users so that they're aware of it.
Specifically, when the satellite connectivity is not available, we need to show users the reason so that
they can act accordingly. For example, if we show that satellite connectivity is not available because
location settings is disabled, users can enable location settings and check availability again.

Test: Manual
Bug: 376694756
Flag: EXEMPT bugfix

Change-Id: I3251e3ee7ead0f1f682999b86b72864f1b6f8dec
Signed-off-by: Duke Lee <deukmo.lee@samsung.corp-partner.google.com>