Ensure TelecomAccountRegistry is HSUM aware.
This CL ensures TelecomAccountRegistry is consistent with HSUM best practices by removing the implication that the primary user is the same as the system user. That assumption is not true in HSUM mode. This CL renames mIsPrimaryUser to be consistent with HSUM best practices.

Test: variable name change only; no functional changes
Bug: 371634365
Flag: EXEMPT variable name update only
Change-Id: I43b648e62a776f45ec2dad0c4b669b73eff57b2c