Add in-service wait timer config for ECC dialed in APM
When emergency call is dialed in airplane mode, telephony must turn
radio power on and start the emergency call. In previous solution, modem
starts network search when radio power is turned on. However, we found
another modem's behavior that it doesn't start network search when
radio power is turned on for emergency call. For this type of modem,
APDS shouldn't wait for modem to get in-service state, and should start
the right away. To support both type of modem, new configuration is
added. If the in-service wait timer value is 0, APDS won't wait for
modem to get in-service.

Flag: EXEMPT bugfix
Bug: 372376726
Test: atest PhoneGlobalsTest TelephonyConnectionServiceTest
Test: Manual test
1. Dialing ECC in APM.
2. Verify that APDS waits for modem to get in-service state up to the
   time configured.
Change-Id: I190d60fe3026eade1d4d1f0e2d92201ce42f9c81